SERVER-49371 Fix a signed/unsigned comparison compile error

+++ src/mongo/util/out_of_line_executor_test.cpp
@@ -122,7 +122,7 @@ TEST(ExecutorTest, InlineRecursiveCountingExecutor) {
             ASSERT(!std::exchange(inTask, true));
             ASSERT_OK(status);
 
-            auto tasksRun = execB->tasksRun.load();
+            auto tasksRun = size_t(execB->tasksRun.load());
             ASSERT_EQ(tasksRun, ++i);
             if (tasksRun < kCountB) {
                 execB->schedule(recurseExec);
